só exibe processos com trafego de rede ativa

+++ timer.h
@@ -6,9 +6,9 @@
 
 // retorna a diferença entre o tempo atual em segundos
 // e o valor em segundos passado por parametro
-float timer(const float old_time);
+double timer(const float old_time);
 
 // retorna a tempo atual em segundos
-float start_timer(void);
+double start_timer(void);
 
 #endif // MTIME_H
